Combine authFetchers into object for easy testing
