+++ ui/src/authFetchers.js
@@ -1,5 +1,21 @@
 import env from "./env";
 
+const VERIFY_TOKEN_MUTATION = `
+  mutation VerifyToken {
+    verifyToken {
+      payload
+    }
+  }
+`;
+
+const verifyToken = () =>
+  fetch(env.GRAPHQL_API_URL, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    credentials: "include",
+    body: JSON.stringify({ query: VERIFY_TOKEN_MUTATION }),
+  });
+
 const REFRESH_TOKEN_MUTATION = `
   mutation RefreshToken {
     refreshToken {
@@ -18,4 +34,7 @@ const refreshToken = () =>
     body: JSON.stringify({ query: REFRESH_TOKEN_MUTATION }),
   });
 
-export default refreshToken;
+export default {
+  verifyToken,
+  refreshToken,
+};
